Update Dockerfile
Some of these are redundant and already listed in the pangeo image, but I didn't have time to scan them all one by one - this is my "canonical" package list from my default `conda-setup` home script. We might as well have all of these packages in...

I also put in some of the config and debian packages from my previous course, based on my experience with having a good setup for students to do 100% of their work on the hub.